+++ ViciEngine/entity.cpp
@@ -1,6 +1,6 @@
 #include "Entity.hpp"
 
-void Entities::Entity::setPosition(double x, double y) {
+void Entities::Entity::setPosition(int x, int y) {
 	_x = x;
 	_y = y;
 }
